Fix a bug in restoring destroyed records.
first(id) does not return the first with the id it returns the first N records. The test happened to work because the "id" was always 1 and the very first component happened to be the component we were interested in in the test.

Use find(id) to make sure we're restoring the correct thing.

Note: also update other test that relied on the global count of components to instead look for a relative delta.